backend/pyproject.toml
[project]
name = "radic-backend"
version = "0.1.0"
description = "Backend for Radic app"
requires-python = ">=3.11"
authors = [
    { name = "Antigravity", email = "antigravity@example.com" }
]
dependencies = [
    "fastapi>=0.121.3",
    "uvicorn[standard]>=0.38.0",
    "supabase>=2.24.0",
    "loguru>=0.7.3",
    "google-genai>=0.3.0",  # New unified SDK (replaces google-generativeai)
    "python-multipart>=0.0.20",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "email-validator>=2.3.0",
]

[tool.uv]
# Managed Python - uv can install and manage Python versions
managed = true

# Dev dependencies - tools needed only for development
dev-dependencies = [
    "pytest>=8.0.0",
    "black>=24.1.0",
    "ruff>=0.1.15",
    "mypy>=1.8.0",
]

[tool.uv.pip]
# Use the new resolver for better dependency resolution
resolution = "highest"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
